Add worker validate call
